+++ sdk/c4-client.ts
@@ -2,8 +2,8 @@
 // Generated from /openapi.json via src/openapi-sdk-gen.js.
 // Do not edit by hand — re-run `c4 openapi --sdk` to refresh.
 
-// Spec version: 1.10.103
-// Generated at: 2026-05-02T15:21:15.479Z
+// Spec version: 1.10.104
+// Generated at: 2026-05-02T15:28:15.746Z
 
 export interface postAuthLoginBody {
   user: string; /** Username */